Add the small non-extending ladder to the grab-capable parts.
The size is set to 5 because it is close in size to a cubic strut,
especially if you consider that without anything to attach on, the
ladder is just three bent bits of metal. Also, 6 of these ladders
are just about enough to match the big ladder, and that's 30.

It is also necessary to fix the handling of isTrigger or the ladder
stops working once grabbed until save/reload.