Refactor handling of pack/unpack events to minimize KASModuleGrab overhead.
This module is added to many common parts like batteries and RCS ports,
so it makes sense to reduce its overall idle overhead by using events
automatically provided locally via SendMessage instead of global hooks.
The grab module was even handling an event for the KASAddonPointer.

The callbacks also must be made public or protected, or it doesn't
even warn if you shadow the method in a subclass.