Allow using KAS pipes as field substitutes for fuel lines.
When linked, either end of the pipe now has a context menu option
that allows enabling or disabling fuel flow in that direction.

Pipe ends attached from EVA in previous versions have to be grabbed
and reattached to use this feature, or the button won't show up.
If both ends are editor-attached, it should work immediately.