Refactor coupling role delegation
Only do it when the source in charge has been asked to decouple (either
via unlink or via the coupling mode change). Do not react on the
external decoupling since it would likely have a bad circular effects.
If they break the link externally with no respect to KAS, then it's
their problem and we just give up.