Allow target to properly lock
Now targets don't go to rejecting state if they are not compatible with
the linking source. However, they still can share a node of the module
that will eventually link. So allow locking from the available state.
Also, disallow locking from accepting link state, since it would mean
there are two modules of the same type on the same node. Such setup is
an error.